requirements.txt
numpy
pandas
scikit-learn
django
djangorestframework
django-cors-headers
requests